% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bubble_plot.R
\name{bubble_plot}
\alias{bubble_plot}
\title{Visualisation of co-ocurrence of indicators in bubble chart}
\usage{
bubble_plot(database, v1, v2, v3)
}
\arguments{
\item{database}{corresponds to the database where the information is stored}

\item{v1}{corresponds to a indicator that will be plotted against the second variable.
This value should be defined as.factor when plotting}

\item{v2}{corresponds to the second indicator that will be plotted.
This value should be defined as.factor when plotting}

\item{v3}{It is a numerical value that will be used to fill in the size and the color of the bubbles in the chart}
}
\value{

}
\description{
Visualisation of co-ocurrence of indicators in bubble chart
}
