% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_ind.R
\name{count_ind}
\alias{count_ind}
\title{Visualization of group of indicators in scenarios}
\usage{
count_ind(x, t)
}
\arguments{
\item{x}{corresponds to the database where the information is stored}

\item{t}{is the variable in the database that will be used to plot the values,
when defining variable write t = "name_variable"}
}
\value{
plot
}
\description{
basic histogram to visualize how many times an indicator was found in the database
}
\examples{
 

data("ind_scenarios")
count_ind(x=ind_scenarios, t="NCP")



}
