% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ind_scenarios}
\alias{ind_scenarios}
\title{ind_scenarios}
\format{
A data frame with 9 variables:
\describe{
\item{NCP_cat}{character. Overall classification of nature contributions to people categories }
\item{NCP}{character. Indicators of different nature contributions to people (NPCs) according to Harrison et al 2019 }
\item{Rewilding_cat}{character. Overall categories for measuring rewilding progress by Torres et al 2018}
\item{rewilding}{character. Specific indicators for measuring the diferent categories of rewilding progress}
\item{overall.pos}{character. Count of participatory scenarios that mention positive benefits both on NCPs and rewilding indicators}
\item{counteracting_posneg}{character. Count of participatory scenarios that mention positive effects for NCPs but negative for rewilding indicators}
\item{counteract_negpos}{character. Count of participatory scenarios that mention negative effects for NCPs but positivefor rewilding indicators}
\item{sum_counteracting}{character. Total number of participatory scenarios that the effects on both indicators were counteracting to each other}
\item{overall.neg}{character. Count of participatory scenarios that mention negative benefits both on NCPs and rewilding indicators }
}
}
\usage{
ind_scenarios
}
\description{
A dataset that summarizes the main indicators for nature contributions to people and rewilding progress in some participatory scenarios
}
\keyword{datasets}
